Refactor: Entity간 연관관계 수정 Entity 간의 연관관계를 수정해줬다. Resolves: #1
