+++ src/main/java/com/example/taleadventure/domain/block/repository/BlockRepository.java
@@ -1,7 +1,7 @@
 package com.example.taleadventure.domain.block.repository;
 
-import com.example.taleadventure.domain.block.entity.BlockEntity;
+import com.example.taleadventure.domain.block.entity.Block;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlockRepository extends JpaRepository<BlockEntity, Long> {
+public interface BlockRepository extends JpaRepository<Block, Long> {
 }
