+++ src/main/java/com/example/taleadventure/domain/day/repository/DayRepository.java
@@ -1,7 +1,7 @@
 package com.example.taleadventure.domain.day.repository;
 
-import com.example.taleadventure.domain.day.entity.DayEntity;
+import com.example.taleadventure.domain.day.entity.Day;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DayRepository extends JpaRepository<DayEntity, Long> {
+public interface DayRepository extends JpaRepository<Day, Long> {
 }
