+++ src/main/java/com/example/taleadventure/domain/member/enummerate/Gender.java
@@ -1,0 +1,6 @@
+package com.example.taleadventure.domain.member.enummerate;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
