+++ src/main/java/com/example/taleadventure/domain/member/enummerate/Status.java
@@ -1,0 +1,6 @@
+package com.example.taleadventure.domain.member.enummerate;
+
+public enum Status {
+    ACTIVE,
+    INACTIVE
+}
