+++ src/main/java/com/example/taleadventure/domain/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.example.taleadventure.domain.member.repository;
+
+import com.example.taleadventure.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
