+++ src/main/java/com/example/taleadventure/domain/talebook/entity/TaleBook.java
@@ -1,9 +1,12 @@
 package com.example.taleadventure.domain.talebook.entity;
 
 import com.example.taleadventure.domain.talebook.enummerate.Category;
+import com.example.taleadventure.domain.word.entity.Word;
 import lombok.*;
 
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @ToString
 @Builder
@@ -11,7 +14,7 @@ import javax.persistence.*;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class TaleBookEntity {
+public class TaleBook {
     @Id
     @GeneratedValue
     @Column(name = "tale_book_id", nullable = false)
@@ -23,4 +26,12 @@ public class TaleBookEntity {
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private Category category;
+
+    @OneToMany
+    @JoinColumn(name = "tale_book_id")
+    private List<Word> words = new ArrayList<>();
+
+    /*@OneToMany
+    @JoinColumn(name = "day_id")
+    private List<Word> words = new ArrayList<>();*/
 }
