+++ src/main/java/com/example/taleadventure/domain/talebook/repository/TaleBookRepository.java
@@ -1,7 +1,7 @@
 package com.example.taleadventure.domain.talebook.repository;
 
-import com.example.taleadventure.domain.talebook.entity.TaleBookEntity;
+import com.example.taleadventure.domain.talebook.entity.TaleBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaleBookRepository extends JpaRepository<TaleBookEntity, Long> {
+public interface TaleBookRepository extends JpaRepository<TaleBook, Long> {
 }
