--- src/main/java/com/example/taleadventure/domain/user/enummerate/Gender.java
@@ -1,6 +1,0 @@
-package com.example.taleadventure.domain.user.enummerate;
-
-public enum Gender {
-    MALE,
-    FEMALE
-}
