--- src/main/java/com/example/taleadventure/domain/user/enummerate/Status.java
@@ -1,6 +1,0 @@
-package com.example.taleadventure.domain.user.enummerate;
-
-public enum Status {
-    ACTIVE,
-    INACTIVE
-}
