--- src/main/java/com/example/taleadventure/domain/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.example.taleadventure.domain.user.repository;
-
-import com.example.taleadventure.domain.user.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-}
