+++ src/main/java/com/example/taleadventure/domain/word/repository/WordRepository.java
@@ -1,7 +1,7 @@
 package com.example.taleadventure.domain.word.repository;
 
-import com.example.taleadventure.domain.word.entity.WordEntity;
+import com.example.taleadventure.domain.word.entity.Word;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WordRepository extends JpaRepository<WordEntity, Long> {
+public interface WordRepository extends JpaRepository<Word, Long> {
 }
