+++ src/main/java/com/example/taleadventure/domain/wordbook/entity/WordBook.java
@@ -1,7 +1,7 @@
 package com.example.taleadventure.domain.wordbook.entity;
 
-import com.example.taleadventure.domain.user.entity.UserEntity;
-import com.example.taleadventure.domain.word.entity.WordEntity;
+import com.example.taleadventure.domain.member.entity.Member;
+import com.example.taleadventure.domain.word.entity.Word;
 import lombok.*;
 
 import javax.persistence.*;
@@ -14,16 +14,16 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class WordBookEntity {
+public class WordBook {
     @Id
     @GeneratedValue
     @Column(name = "word_book_id", nullable = false)
     private Long id;
 
-    @OneToOne
-    private UserEntity userEntity;
+    /*@OneToOne
+    private Member member;*/
 
     @OneToMany
     @JoinColumn(name = "word_book_id")
-    private List<WordEntity> wordEntities = new ArrayList<>();
+    private List<Word> words = new ArrayList<>();
 }
