+++ src/main/java/com/example/taleadventure/domain/wordbook/repository/WordBookRepository.java
@@ -1,7 +1,7 @@
 package com.example.taleadventure.domain.wordbook.repository;
 
-import com.example.taleadventure.domain.wordbook.entity.WordBookEntity;
+import com.example.taleadventure.domain.wordbook.entity.WordBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WordBookRepository extends JpaRepository<WordBookEntity, Long> {
+public interface WordBookRepository extends JpaRepository<WordBook, Long> {
 }
